FIX: CLI location
