--- cli/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	cliutils "github.com/fredyk/westack-go/cli-utils"
-)
-
-func main() {
-	cliutils.RunCli()
-}
