+++ v2/tests/westack_test.go
@@ -5,6 +5,7 @@ import (
 	"crypto/rand"
 	"encoding/json"
 	"fmt"
+	"github.com/fredyk/westack-go/v2/westack"
 	"io"
 	"log"
 	"math/big"
@@ -22,7 +23,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/fredyk/westack-go/v2"
 	wst "github.com/fredyk/westack-go/v2/common"
 )
